Tests and refactoring observer loop
I found that break from a select stays in the loop, but a default
block doesn't see to always finish before the next select when there
is a sleep. This were sorted out by moving the exit to a channel that
can be included in the select and refactoring the Observer.Close
function to be a noop when it detects the observer has already finished.